Oct 14 2025
Updates to the Prescription component; corrections made for JWT service
and a warning to ensure that MySQL Server is in fact running, and a
recommendation to set the service to start automatically. The other
error was the accidental deletion of JWT token from
application.properties. Both of these issues will create compilation
errors.